Share secp256k1 test helpers across crypto tests
The secp256k1 group order `SECP256K1_N`, the high-S flip helper, and the
P2PKH lock-script builder were each duplicated across the transaction,
messages and script test modules.

Add a `#[cfg(test)] test_util` module with a single `SECP256K1_N`
constant and `flip_to_high_s` helper, and route the high-S tests in
transaction/mod.rs, messages/tx_chronicle_validate.rs and
script/checker.rs through it. The hand-rolled P2PKH lock scripts in the
tx_chronicle_validate and checker tests now call the existing
`p2pkh::create_lock_script`. (The p2pkh.rs test that hand-builds a script
to exercise `check_lock_script` is left as-is, deliberately independent.)

Test-only change; cargo test --all-features passes (194 + 10 doctests),
no warnings.
