+++ src/lib.rs
@@ -25,3 +25,6 @@ pub mod interface;
 /// Python (PyO3) bindings, compiled only with the `python` feature.
 #[cfg(feature = "python")]
 pub mod python;
+
+#[cfg(test)]
+mod test_util;
